compilation warnings fix
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15858 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/asn/asn_val.c
@@ -985,8 +985,8 @@ asn_impl_read_value_field(const asn_value *container,  void *data,
                           size_t *d_len, char *field_labels)
 {
     const asn_value *value;
-    int m_len;
-    int rc;
+    size_t           m_len;
+    int              rc;
 
     if (!container) return ETEWRONGPTR; 
 
@@ -1048,7 +1048,7 @@ asn_impl_read_value_field(const asn_value *container,  void *data,
                 return ETESMALLBUF;
 
             *d_len = value->len;
-            memcpy (data, value->data.other, m_len);
+            memcpy(data, value->data.other, m_len);
         }
         break;
